# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Interacción con Base de Datos FireBird",
        "description": "API para la gestión",
        "version": "1.0"
    },
    "paths": {
...
        "/api/inventarios": {
            "get": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Obtener todos los inventarios",
                "operationId": "721655bcda437b9623a493da77e2fd43",
                "responses": {
                    "200": {
                        "description": "Lista de inventarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Inventario"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Crear un nuevo inventario",
                "operationId": "e2544a669fb3cce9342d9cf42dfbf205",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Inventario"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inventario creado exitosamente"
                    }
                }
            }
        },
        "/api/inventarios/{id}": {
            "get": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Obtener un inventario por ID",
                "operationId": "61ca5382ad56e1c47831198872094a4a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del inventario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del inventario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Inventario"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Actualizar un inventario",
                "operationId": "dd96c5e900591f5e268aa18a8320e524",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del inventario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Inventario"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inventario actualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Eliminar un inventario",
                "operationId": "72192e1c345bc6be025c87d0b8b860d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del inventario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Inventario eliminado"
                    }
                }
            }
        },
...
        "/api/inventarios/restore/{id}": {
            "put": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Restaurar un inventario eliminado",
                "operationId": "d395dac2b6ab40c5ba3b4739a1cdda3d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del inventario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inventario restaurado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Inventario": {
                "title": "Inventario",
                "description": "Modelo de Inventario",
                "required": [
                    "NOMBRE",
                    "ID_FOTOS",
                    "ID_USUARIOS"
                ],
                "properties": {
                    "NOMBRE": {
                        "description": "Nombre del inventario",
                        "type": "string"
                    },
                    "ID_FOTOS": {
                        "description": "ID de la foto asociada",
                        "type": "integer"
                    },
                    "ID_USUARIOS": {
                        "description": "ID del usuario propietario",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}